vertex-level_regression/requirements.txt
absl-py==0.9.0
alabaster==0.7.10
asn1crypto==0.24.0
astor==0.8.1
astroid==1.6.1
astropy==2.0.3
attrs==17.4.0
Babel==2.5.3
backports.shutil-get-terminal-size==1.0.0
beautifulsoup4==4.6.0
bitarray==0.8.1
bkcharts==0.2
blaze==0.11.3
bleach==2.1.2
bokeh==0.12.13
boto==2.48.0
Bottleneck==1.2.1
certifi==2020.4.5.1
cffi==1.12.3
chardet==3.0.4
chumpy==0.69
click==6.7
cloudpickle==0.5.2
clyent==1.2.2
colorama==0.3.9
cycler==0.10.0
Cython==0.27.3
cytoolz==0.9.0
dask==0.16.1
datashape==0.5.4
decorator==4.2.1
distributed==1.20.2
docutils==0.14
glob2==0.6
gmpy2==2.0.8
greenlet==0.4.12
grpcio==1.27.2
h5py==2.10.0
heapdict==1.0.0
html5lib==1.0.1
idna==2.6
imageio==2.2.0
imagesize==0.7.1
ipykernel==4.8.0
ipython==6.2.1
ipython-genutils==0.2.0
ipywidgets==7.1.1
isort==4.2.15
itsdangerous==0.24
joblib==0.13.2
jsonpatch==1.32
jsonpointer==2.1
jsonschema==2.6.0
Keras-Applications==1.0.8
Keras-Preprocessing==1.1.0
llvmlite==0.21.0
locket==0.2.0
lxml==4.1.1
Mako==1.1.3
Markdown==3.2.1
MarkupSafe==1.0
matplotlib==2.1.2
mccabe==0.6.1
mistune==0.8.3
mock==4.0.1
mpmath==1.0.0
msgpack==1.0.0
msgpack-numpy==0.4.6.post0
msgpack-python==0.5.1
multipledispatch==0.4.9
natsort==7.0.1
navigator-updater==0.1.0
nbconvert==5.3.1
nbformat==4.4.0
networkx==2.1
nltk==3.2.5
nose==1.3.7
notebook==5.4.0
numba==0.36.2
numexpr==2.6.4
numpngw==0.0.8
numpy==1.15.4
numpydoc==0.7.0
odo==0.5.1
olefile==0.46
OpenEXR==1.3.2
openpyxl==2.4.10
packaging==16.8
pandas==0.22.0
pandocfilters==1.4.2
parso==0.1.1
partd==0.3.8
path.py==10.5
pathlib2==2.3.0
patsy==0.5.0
pep8==1.7.1
pexpect==4.3.1
pickle5==0.0.11
pickleshare==0.7.4
Pillow==5.4.1
pkginfo==1.4.1
pluggy==0.6.0
ply==3.10
prompt-toolkit==1.0.15
protobuf==3.11.3
psutil==5.4.3
ptyprocess==0.5.2
py==1.5.2
pycocotools==2.0.1
pycodestyle==2.3.1
pycosat==0.6.3
pycparser==2.19
pycrypto==2.6.1
pycuda==2019.1.2
pycurl==7.43.0.1
pyflakes==1.6.0
Pygments==2.2.0
pylint==1.8.2
pymesh==1.0.2
pyodbc==4.0.22
pyOpenSSL==17.5.0
pyparsing==2.2.0
pypoisson==0.10
PySocks==1.6.7
pytest==3.3.2
python-dateutil==2.6.1
pytools==2020.3
pytz==2017.3
PyWavelets==0.5.2
PyYAML==3.12
pyzmq==16.0.3
QtAwesome==0.4.4
qtconsole==4.3.1
QtPy==1.3.1
requests==2.18.4
rope==0.10.7
ruamel-yaml==0.15.35
scikit-image==0.13.1
scikit-learn==0.21.3
scipy==1.2.0
six==1.12.0
snowballstemmer==1.2.1
sortedcollections==0.5.3
sortedcontainers==1.5.9
statsmodels==0.8.0
sympy==1.1.1
tables==3.4.2
tabulate==0.8.7
tblib==1.3.2
tensorboard==1.13.1
tensorflow-estimator==1.13.0
tensorflow-gpu==1.13.1
tensorflow-graphics==1.0.0
tensorpack==0.9.0
termcolor==1.1.0
terminado==0.8.1
testpath==0.3.1
toolz==0.9.0
tornado==4.5.3
tqdm==4.46.1
traitlets==4.3.2
trimesh==3.9.18
typing==3.6.2
unicodecsv==0.14.1
urllib3==1.22
visdom==0.1.8.9
wcwidth==0.1.7
webencodings==0.5.1
websocket-client==0.59.0
Werkzeug==0.14.1
widgetsnbextension==3.1.0
wrapt==1.10.11
xlrd==1.1.0
XlsxWriter==1.0.2
xlwt==1.3.0
zict==0.1.3
